cli/requirements_cli.txt
# Proteus AI CLI工具依赖包

# HTTP客户端
aiohttp>=3.8.0
requests>=2.28.0

# SSE客户端
sseclient-py>=1.7.0

# 终端颜色支持
colorama>=0.4.4

# 命令行参数解析 (Python标准库，无需安装)
# argparse

# JSON处理 (Python标准库，无需安装)
# json

# 异步支持 (Python 3.7+标准库，无需安装)
# asyncio